Remove Kramdown overrrides
Currently, in order to disable ordered lists for legislative lists, we modify
the @@parsers class variable and the LIST_START constant of the Kramdown
library.

We are seeing a bunch of errors relating to these missing in production, such
as:
- NoMethodError: undefined method `start_re' for nil:NilClass (NoMethodError)
  if @src.check(@parsers[name].start_re)
- Kramdown::Error: Unknown parser: list (Kramdown::Error)
- NameError: uninitialized constant Kramdown::Parser::Kramdown::LIST_START

Rather than modifying class variables and constants in a non thread-safe
manner, this takes the simpler approch of copying the body of the `parse_list`
method and making the neccessary changes required for disabling ordered lists.

The downside of copying the majority body of the `parse_list` method may be
offset by the fact that:
- This is thread-safe
- This is significantly simpler to understand
- The list module of Kramdown is rarely updated.
- The previous implementation also relied on the internals of Kramdown.