Fix Rubocop Style/HashEachMethods
We have a false positive here from the unsafe Style/HashEachMethods cop.
Rubocop thinks that we are calling `each` on a hash, but we are actually
calling it on an array of two-element arrays.

By putting parentheses around the block arguments we can indicate that we're
not working with a hash, and suppress RuboCop offenses. See
https://docs.rubocop.org/rubocop/cops_style.html#stylehasheachmethods