Fix single line formatting of call to action component
We use the default mechanism for parsing the call to action component so that
the parser handles abbreviations and footnotes within.

Since this change (#292) we have seen a regression where defining a call to
action on a single line does not work as expected (e.g., `$CTA Click here to
start the program $CTA`). The closing `div` is evaluated as markdown and the
call to action component therefore runs on until the next closing `div`.

This can be fixed by adding in some new lines between the div tags.