Pin version of `govuk_publishing_components`
There was a change in version 38.4.2 of `govuk_publishing_components`
which resulted in additional tag attributes being added for some
elements.

We therefore have test failures because of this.

Pinning the version to one which allows the tests to continue passing.
Dependabot will now raise a PR to bump the version of
`govuk_publishing_components`, which will allow us to fix the failing
tests separately.