Support embeds in Govspeak
This adds a new `embeds` extenstion, which scans the body for the embed
regex and uses the EmbedPresenter to return the appropriate text,
depending on the content type, assuming an array of editions has been
passed to the `embeds` argument.

For now, we’re reimplementing the logic already in the Publishing API
(along with some additional classes, IDs etc), but in future we can
expand the presenter to support more complex logic.