Remove ContentBlock related functionality
Because of a degree of complexity in how Govspeak is rendered in the
Publishing Platform, we can’t always fetch embedded content at the
point where Govspeak is rendered (this is particuarly the case in
Whitehall). With this in mind, and after some discussion with the team,
we’ve decided that, for now, it’s better to handle embeds in the
Publishing API.

With this in mind, rather than letting this functionality die on the
vine within Govspeak, we should remove the functionality altogether. As
this is pretty new code (and the content modelling team are the only
people aware of it and likely to use it), we should remove it from the
codebase entirely.