Allow data attributes in span tags
The work we’re doing in Content Modelling means we will be rendering
content blocks within span tags with data attributes, which will make
identifying content blocks easier. As we may be re-presenting content
block embeds to Govspeak, we need to ensure that these tags don’t get
stripped out, so we need to add `data-*` attributes to the allowlist for
span tags.