Update Kramdown role attribute placement
An update to Kramdown to support versions 2.5.x produced 14 failing
tests, stemming from:
- [a change to the placement of the doc-noteref role
  attribute](https://github.com/gettalong/kramdown/blob/fc051a9d93e4dc3ff05bf41b70a79297ebdb669f/doc/news/release_2_5_0.page#L38-L39)
- [the removal of the doc-endnote role
  attribute](https://github.com/gettalong/kramdown/blob/fc051a9d93e4dc3ff05bf41b70a79297ebdb669f/doc/news/release_2_5_0.page#L30)

This fixes our expectations to match the new behaviour. We don't
customise any of this in Govspeak, so updating the tests to continue to
match the library default seems reasonable

Because this is a gem and therefore uses a gemspec rather than a
lockfile, running `bundle install` locally doesn't necessarily bring you
in line with the versions the CI tests are running on. This is
presumably because already-installed versions of dependencies might
satisfy the ranges allowed by the gemspec. I therefore wasn't seeing
these errors after a `bundle install`, since I had Kramdown 2.4.0
installed. Running `bundle update`, meanwhile, will presumably bring in the latest
allowed versions. After doing this, I saw the errors

It's worth noting then that the specs will be very slightly inaccurate -
in terms of what they assert about library defaults - for dependents
running Kramdown 2.3.x or 2.4.x. We could update the minimum required
Kramdown version to 2.5.x to address this, but it feels like a fairly
minor issue, and as mentioned earlier, we don't customise any of this in
Govspeak. It might be worth exploring whether we can reduce the extent
to which we assert on library-default versus Govspeak-customised output
in future work to simplify Govspeak testing. We'll need to consider how
precise Govspeak needs to be about its output as a gem