Break up long string into a heredoc
I know heredocs are weird, but this is *probably* more readable.

I always need to remind myself how the various flavours of heredoc
actually behave. Squiggly heredocs (<<~) remove indentation from their
content. It'll remove the same amount of indentation across all lines,
equal to the indentation of its least-indented line. In this case, all
of our lines start out indented equally, so all indentation ends up
removed from the resulting string.

Also worth noting: the replacement of newline chars in the old string
with actual line breaks in the heredoc. The tests pass, so I hope that
that means the output is identical in all cases that we care about.