Make line break removal more targetted
This `sub` was present in the very first commit that introduced the
address block extension and there's no explanation of what it's
specifically intended to achieve.

I've interpreted it as "if there's a leading line break, remove it", as
opposed to "remove the first occurrence of a line break regardless of
where it appears in the input string". I could have missed something.

This change makes the intention that I've assumed was there clearer to
readers and prevents any other line break being removed unintentionally
if something did ever change about the expected format of the input.