Drop support for Ruby <3.2
We're looking into compatibility issues with a recent release of the
govuk_publishing_components gem. Unfortunately, since the components gem
dropped support for Ruby <3.2 in its own 41.1.0 release (August 2024),
we need to use a newer version of Ruby to be able to actually reproduce
the problem.

While updating the .ruby-version (to 3.2.2, same as
govuk_publishing_components), it seems sensible to update the minimum
supported Ruby version to match it.