Add a minimalist html-parsing test helper
In some tests we're using string matching/regex to make assertions about
rendered HTML. Those assertions can be overly sensitive to, say, the
order that an HTML element's attributes are listed in.

This is particularly an issue where Govspeak invokes
GovukPublishingComponents, since we no longer pin the version of that
dependency and different versions can produce different HTML.

Parsing the HTML string and making assertions against an object
representation is likely to be more reliable.

---

For this method, I've copied an approach taken in one of the assertions
provided by the existing GovspeakTestHelper. GovspeakTestHelper is
designed in such a way that our tests that make use of it need to be
written in quite a rigid format. It has been suggested once in a while
that we revisit the design of this gem's tests and test helpers
generally. For now, this new method is standalone and flexible enough to
just insert into our more vanilla tests.