Parse HTML in tests instead of using regex
This isn't an exhaustive update, I'm only changing the ones that're
making the CI build fail right now.

A change in the order of an HTML element's attributes has caused the
build to start failing. The regex pattern expects the class name to
immediately follow the tag name, but that's no longer true -- due to
a change in our (unpinned) govuk_publishing_components dependency.

This new assertion works with older and newer versions of the components
gem.