Allow acronyms/abbreviations in steps
Abbreviations were not working when embedded in a step.

Fixing this in the same way as abbreviations in legislative lists:
wrapping the list item in an element that has a `markdown="1"` value
defined, then indenting the contents correctly so that the markdown is
parsed again during the second round of processing.

Note: this affects the indentation, so the existing tests are being
updated to reflect that.