Revert "Add GA4 indexes to attachments that render a details component"
This reverts commit ecdda77d0f39e115ae3b4d2a73f6fd0fd94e915f.

The code was working, but it isn't giving us exactly what we want - it was returning all attachments for a given document, but we want to be able to filter by ones which are actually being rendered on the page. Therefore we'll explore using JS to handle this, or shelve this work.