WIP expose to_hash_ast method
The transform_ast is very work in progress (i.e. basically not
implemented), but the idea would be to transform kramdown's
representation into something more standard (like mdast).

At least converting the govspeak extensions that are parsed as HTML by
kramdown into something more meaningful is probably necessary.

I had a bit of a play trying to use pattern matching for the tests. I'm
not sure it's a great approach though - failures tend to be pretty hard
to read.