Localise image credit label where possible
... which now includes Welsh

The `ImagePresenter` code looked like it was handling translations, but
it wasn't. We don't set `I18n.locale` to anything other than English in
the Govspeak code, so we need to pass the locale explicitly when calling
`I18n.t`. We already do this for contacts via the `TemplateRender#t`
wrapper