WIP parser and converter for Informational
The goal here is producing a more concise AST entry from the parser and
maintaining the same HTML output as the `extension` version.

A note about rewriting PARAGRAPH_END:

According to our tests, Govspeak currently supports an Informational
being declared immediately following a line of paragraph text, separated
by a single line break. If I've understood it correctly, Kramdown wants there to be 2 line breaks between blocks.

In the long run, we could update our guide and let users know that
blocks should have 2 line breaks between them. For now, if we are going
to migrate our preprocessing `extension`s to Kramdown parsers, it'd be
nice not to change external stuff in the process.

So, we've modified the PARAGRAPH_END pattern because in order to support
those single line breaks, Kramdown's paragraph parser needs to recognise
the beginning of an Informational as a valid marker of the end of
a paragraph.