Ensure i18n load paths are flattened
In Rails >= 8.1, there’s been a change that means the i18n load path
is expected to be a flat array. Bringing Govspeak into apps that use
Rails >= 8.1 gives the following error:

```
Failure/Error: Rails.application.initialize!

TypeError:
  no implicit conversion of Array into String
snip
```

Adding a splat when setting the `locale_paths` variable has, no effect
so we’re just adding the `locale_paths` array to the `I18n.load_path`
as one item. Moving the splat operator to the `unshift` call ensures
the array is destructed as we expect.

I’ve added a test to confirm this is the case too.