Remove pinning of `minitest` version
The version of `minitest` was pinned to 5.8 in
2aceaecdb1d2f215338e30aa55877ce945a8163a.

Later on, in 9fe1b3dfe2ec8c4b81fa07620097fdf78527bc29 we loosened the
pin to only the major version.

However `minitest` version 6 has now been released and this restriction
prevents that upgrade from occuring in apps that use `govspeak`.

There appears to be no reason why we originally pinned the gem to this
version, so removing the version completely. This will allow apps using
this gem to upgrade `minitest` in their own time, using either 5 or 6
until they wish to upgrade.